+++ src/pip/_internal/distributions/sdist.py
